book_gen_system/requirements.txt
supabase
python-docx
google-generativeai
pandas
openpyxl
fpdf2
requests
python-dotenv
duckduckgo-search
openai
